chore: move registering into package file
