+++ registry/checksum/checksum.go
@@ -20,3 +20,11 @@ func (cr *ChecksumRegistry) Uid() string {
 func (cr *ChecksumRegistry) LinkHandler(fh sprout.Handler) {
 	cr.handler = &fh
 }
+
+// RegisterFunctions registers all functions of the registry.
+func (cr *ChecksumRegistry) RegisterFunctions(funcsMap sprout.FunctionMap) {
+	sprout.AddFunction(funcsMap, "sha1sum", cr.Sha1sum)
+	sprout.AddFunction(funcsMap, "sha256sum", cr.Sha256sum)
+	sprout.AddFunction(funcsMap, "adler32sum", cr.Adler32sum)
+	sprout.AddFunction(funcsMap, "md5sum", cr.Md5sum)
+}
