+++ registry/conversion/functions.go
@@ -5,23 +5,8 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/go-sprout/sprout"
 	"github.com/spf13/cast"
 )
-
-func (cr *ConversionRegistry) RegisterFunctions(funcsMap sprout.FunctionMap) {
-	sprout.AddFunction(funcsMap, "toBool", cr.ToBool)
-	sprout.AddFunction(funcsMap, "toInt", cr.ToInt)
-	sprout.AddFunction(funcsMap, "toInt64", cr.ToInt64)
-	sprout.AddFunction(funcsMap, "toUint", cr.ToUint)
-	sprout.AddFunction(funcsMap, "toUint64", cr.ToUint64)
-	sprout.AddFunction(funcsMap, "toFloat64", cr.ToFloat64)
-	sprout.AddFunction(funcsMap, "toOctal", cr.ToOctal)
-	sprout.AddFunction(funcsMap, "toString", cr.ToString)
-	sprout.AddFunction(funcsMap, "toDate", cr.ToDate)
-	sprout.AddFunction(funcsMap, "toDuration", cr.ToDuration)
-	sprout.AddFunction(funcsMap, "mustToDate", cr.MustToDate)
-}
 
 // ToBool converts a value to a boolean.
 //
