+++ registry/crypto/crypto.go
@@ -62,3 +62,20 @@ func (ch *CryptoRegistry) Uid() string {
 func (ch *CryptoRegistry) LinkHandler(fh sprout.Handler) {
 	ch.handler = &fh
 }
+
+// RegisterFunctions adds all crypto-related functions to the provided registry.
+func (ch *CryptoRegistry) RegisterFunctions(funcsMap sprout.FunctionMap) {
+	sprout.AddFunction(funcsMap, "bcrypt", ch.Bcrypt)
+	sprout.AddFunction(funcsMap, "htpasswd", ch.Htpasswd)
+	sprout.AddFunction(funcsMap, "derivePassword", ch.DerivePassword)
+	sprout.AddFunction(funcsMap, "genPrivateKey", ch.GeneratePrivateKey)
+	sprout.AddFunction(funcsMap, "buildCustomCert", ch.BuildCustomCertificate)
+	sprout.AddFunction(funcsMap, "genCA", ch.GenerateCertificateAuthority)
+	sprout.AddFunction(funcsMap, "genCAWithKey", ch.GenerateCertificateAuthorityWithPEMKey)
+	sprout.AddFunction(funcsMap, "genSelfSignedCert", ch.GenerateSelfSignedCertificate)
+	sprout.AddFunction(funcsMap, "genSelfSignedCertWithKey", ch.GenerateSelfSignedCertificateWithPEMKey)
+	sprout.AddFunction(funcsMap, "genSignedCert", ch.GenerateSignedCertificate)
+	sprout.AddFunction(funcsMap, "genSignedCertWithKey", ch.GenerateSignedCertificateWithPEMKey)
+	sprout.AddFunction(funcsMap, "encryptAES", ch.EncryptAES)
+	sprout.AddFunction(funcsMap, "decryptAES", ch.DecryptAES)
+}
