+++ registry/maps/maps.go
@@ -20,3 +20,22 @@ func (mr *MapsRegistry) Uid() string {
 func (mr *MapsRegistry) LinkHandler(fh sprout.Handler) {
 	mr.handler = &fh
 }
+
+// RegisterFunctions registers all functions of the registry.
+func (mr *MapsRegistry) RegisterFunctions(funcsMap sprout.FunctionMap) {
+	sprout.AddFunction(funcsMap, "dict", mr.Dict)
+	sprout.AddFunction(funcsMap, "get", mr.Get)
+	sprout.AddFunction(funcsMap, "set", mr.Set)
+	sprout.AddFunction(funcsMap, "unset", mr.Unset)
+	sprout.AddFunction(funcsMap, "keys", mr.Keys)
+	sprout.AddFunction(funcsMap, "values", mr.Values)
+	sprout.AddFunction(funcsMap, "pluck", mr.Pluck)
+	sprout.AddFunction(funcsMap, "pick", mr.Pick)
+	sprout.AddFunction(funcsMap, "omit", mr.Omit)
+	sprout.AddFunction(funcsMap, "dig", mr.Dig)
+	sprout.AddFunction(funcsMap, "hasKey", mr.HasKey)
+	sprout.AddFunction(funcsMap, "merge", mr.Merge)
+	sprout.AddFunction(funcsMap, "mergeOverwrite", mr.MergeOverwrite)
+	sprout.AddFunction(funcsMap, "mustMerge", mr.MustMerge)
+	sprout.AddFunction(funcsMap, "mustMergeOverwrite", mr.MustMergeOverwrite)
+}
