+++ registry/reflect/reflect.go
@@ -20,3 +20,15 @@ func (rr *ReflectRegistry) Uid() string {
 func (rr *ReflectRegistry) LinkHandler(fh sprout.Handler) {
 	rr.handler = &fh
 }
+
+// RegisterFunctions registers all functions of the registry.
+func (nr *ReflectRegistry) RegisterFunctions(funcsMap sprout.FunctionMap) {
+	sprout.AddFunction(funcsMap, "typeIs", nr.TypeIs)
+	sprout.AddFunction(funcsMap, "typeIsLike", nr.TypeIsLike)
+	sprout.AddFunction(funcsMap, "typeOf", nr.TypeOf)
+	sprout.AddFunction(funcsMap, "kindIs", nr.KindIs)
+	sprout.AddFunction(funcsMap, "kindOf", nr.KindOf)
+	sprout.AddFunction(funcsMap, "deepEqual", nr.DeepEqual)
+	sprout.AddFunction(funcsMap, "deepCopy", nr.DeepCopy)
+	sprout.AddFunction(funcsMap, "mustDeepCopy", nr.MustDeepCopy)
+}
