+++ registry/regexp/functions.go
@@ -2,26 +2,7 @@ package regexp
 
 import (
 	"regexp"
-
-	"github.com/go-sprout/sprout"
 )
-
-// RegisterFunctions registers all functions of the registry.
-func (rr *RegexpRegistry) RegisterFunctions(funcsMap sprout.FunctionMap) {
-	sprout.AddFunction(funcsMap, "regexFind", rr.RegexFind)
-	sprout.AddFunction(funcsMap, "regexFindAll", rr.RegexFindAll)
-	sprout.AddFunction(funcsMap, "regexMatch", rr.RegexMatch)
-	sprout.AddFunction(funcsMap, "regexSplit", rr.RegexSplit)
-	sprout.AddFunction(funcsMap, "regexReplaceAll", rr.RegexReplaceAll)
-	sprout.AddFunction(funcsMap, "regexReplaceAllLiteral", rr.RegexReplaceAllLiteral)
-	sprout.AddFunction(funcsMap, "regexQuoteMeta", rr.RegexQuoteMeta)
-	sprout.AddFunction(funcsMap, "mustRegexFind", rr.MustRegexFind)
-	sprout.AddFunction(funcsMap, "mustRegexFindAll", rr.MustRegexFindAll)
-	sprout.AddFunction(funcsMap, "mustRegexMatch", rr.MustRegexMatch)
-	sprout.AddFunction(funcsMap, "mustRegexSplit", rr.MustRegexSplit)
-	sprout.AddFunction(funcsMap, "mustRegexReplaceAll", rr.MustRegexReplaceAll)
-	sprout.AddFunction(funcsMap, "mustRegexReplaceAllLiteral", rr.MustRegexReplaceAllLiteral)
-}
 
 // RegexFind returns the first match of the regex pattern in the string.
 //
