+++ registry/regexp/regexp.go
@@ -20,3 +20,20 @@ func (rr *RegexpRegistry) Uid() string {
 func (rr *RegexpRegistry) LinkHandler(fh sprout.Handler) {
 	rr.handler = &fh
 }
+
+// RegisterFunctions registers all functions of the registry.
+func (rr *RegexpRegistry) RegisterFunctions(funcsMap sprout.FunctionMap) {
+	sprout.AddFunction(funcsMap, "regexFind", rr.RegexFind)
+	sprout.AddFunction(funcsMap, "regexFindAll", rr.RegexFindAll)
+	sprout.AddFunction(funcsMap, "regexMatch", rr.RegexMatch)
+	sprout.AddFunction(funcsMap, "regexSplit", rr.RegexSplit)
+	sprout.AddFunction(funcsMap, "regexReplaceAll", rr.RegexReplaceAll)
+	sprout.AddFunction(funcsMap, "regexReplaceAllLiteral", rr.RegexReplaceAllLiteral)
+	sprout.AddFunction(funcsMap, "regexQuoteMeta", rr.RegexQuoteMeta)
+	sprout.AddFunction(funcsMap, "mustRegexFind", rr.MustRegexFind)
+	sprout.AddFunction(funcsMap, "mustRegexFindAll", rr.MustRegexFindAll)
+	sprout.AddFunction(funcsMap, "mustRegexMatch", rr.MustRegexMatch)
+	sprout.AddFunction(funcsMap, "mustRegexSplit", rr.MustRegexSplit)
+	sprout.AddFunction(funcsMap, "mustRegexReplaceAll", rr.MustRegexReplaceAll)
+	sprout.AddFunction(funcsMap, "mustRegexReplaceAllLiteral", rr.MustRegexReplaceAllLiteral)
+}
