+++ registry/slices/functions.go
@@ -7,47 +7,9 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/go-sprout/sprout"
 	"github.com/go-sprout/sprout/internal/helpers"
 	"github.com/spf13/cast"
 )
-
-// RegisterFunctions registers all functions of the registry.
-func (sr *SlicesRegistry) RegisterFunctions(funcsMap sprout.FunctionMap) {
-	sprout.AddFunction(funcsMap, "list", sr.List)
-	sprout.AddFunction(funcsMap, "append", sr.Append)
-	sprout.AddFunction(funcsMap, "prepend", sr.Prepend)
-	sprout.AddFunction(funcsMap, "concat", sr.Concat)
-	sprout.AddFunction(funcsMap, "chunk", sr.Chunk)
-	sprout.AddFunction(funcsMap, "uniq", sr.Uniq)
-	sprout.AddFunction(funcsMap, "compact", sr.Compact)
-	sprout.AddFunction(funcsMap, "slice", sr.Slice)
-	sprout.AddFunction(funcsMap, "has", sr.Has)
-	sprout.AddFunction(funcsMap, "without", sr.Without)
-	sprout.AddFunction(funcsMap, "rest", sr.Rest)
-	sprout.AddFunction(funcsMap, "initial", sr.Initial)
-	sprout.AddFunction(funcsMap, "first", sr.First)
-	sprout.AddFunction(funcsMap, "last", sr.Last)
-	sprout.AddFunction(funcsMap, "reverse", sr.Reverse)
-	sprout.AddFunction(funcsMap, "sortAlpha", sr.SortAlpha)
-	sprout.AddFunction(funcsMap, "splitList", sr.SplitList)
-	sprout.AddFunction(funcsMap, "strSlice", sr.StrSlice)
-	sprout.AddFunction(funcsMap, "until", sr.Until)
-	sprout.AddFunction(funcsMap, "untilStep", sr.UntilStep)
-	sprout.AddFunction(funcsMap, "mustAppend", sr.MustAppend)
-	sprout.AddFunction(funcsMap, "mustPrepend", sr.MustPrepend)
-	sprout.AddFunction(funcsMap, "mustChunk", sr.MustChunk)
-	sprout.AddFunction(funcsMap, "mustUniq", sr.MustUniq)
-	sprout.AddFunction(funcsMap, "mustCompact", sr.MustCompact)
-	sprout.AddFunction(funcsMap, "mustSlice", sr.MustSlice)
-	sprout.AddFunction(funcsMap, "mustHas", sr.MustHas)
-	sprout.AddFunction(funcsMap, "mustWithout", sr.MustWithout)
-	sprout.AddFunction(funcsMap, "mustRest", sr.MustRest)
-	sprout.AddFunction(funcsMap, "mustInitial", sr.MustInitial)
-	sprout.AddFunction(funcsMap, "mustFirst", sr.MustFirst)
-	sprout.AddFunction(funcsMap, "mustLast", sr.MustLast)
-	sprout.AddFunction(funcsMap, "mustReverse", sr.MustReverse)
-}
 
 // List creates a list from the provided elements.
 //
