+++ registry/strings/strings.go
@@ -90,3 +90,46 @@ func (sr *StringsRegistry) Uid() string {
 func (sr *StringsRegistry) LinkHandler(fh sprout.Handler) {
 	sr.handler = &fh
 }
+
+// RegisterFunctions registers all functions of the registry.
+func (sr *StringsRegistry) RegisterFunctions(funcsMap sprout.FunctionMap) {
+	sprout.AddFunction(funcsMap, "nospace", sr.Nospace)
+	sprout.AddFunction(funcsMap, "trim", sr.Trim)
+	sprout.AddFunction(funcsMap, "trimAll", sr.TrimAll)
+	sprout.AddFunction(funcsMap, "trimPrefix", sr.TrimPrefix)
+	sprout.AddFunction(funcsMap, "trimSuffix", sr.TrimSuffix)
+	sprout.AddFunction(funcsMap, "contains", sr.Contains)
+	sprout.AddFunction(funcsMap, "hasPrefix", sr.HasPrefix)
+	sprout.AddFunction(funcsMap, "hasSuffix", sr.HasSuffix)
+	sprout.AddFunction(funcsMap, "toLower", sr.ToLower)
+	sprout.AddFunction(funcsMap, "toUpper", sr.ToUpper)
+	sprout.AddFunction(funcsMap, "replace", sr.Replace)
+	sprout.AddFunction(funcsMap, "repeat", sr.Repeat)
+	sprout.AddFunction(funcsMap, "join", sr.Join)
+	sprout.AddFunction(funcsMap, "trunc", sr.Trunc)
+	sprout.AddFunction(funcsMap, "shuffle", sr.Shuffle)
+	sprout.AddFunction(funcsMap, "ellipsis", sr.Ellipsis)
+	sprout.AddFunction(funcsMap, "ellipsisBoth", sr.EllipsisBoth)
+	sprout.AddFunction(funcsMap, "initials", sr.Initials)
+	sprout.AddFunction(funcsMap, "plural", sr.Plural)
+	sprout.AddFunction(funcsMap, "wrap", sr.Wrap)
+	sprout.AddFunction(funcsMap, "wrapWith", sr.WrapWith)
+	sprout.AddFunction(funcsMap, "quote", sr.Quote)
+	sprout.AddFunction(funcsMap, "squote", sr.Squote)
+	sprout.AddFunction(funcsMap, "toCamelCase", sr.ToCamelCase)
+	sprout.AddFunction(funcsMap, "toKebabCase", sr.ToKebabCase)
+	sprout.AddFunction(funcsMap, "toPascalCase", sr.ToPascalCase)
+	sprout.AddFunction(funcsMap, "toDotCase", sr.ToDotCase)
+	sprout.AddFunction(funcsMap, "toPathCase", sr.ToPathCase)
+	sprout.AddFunction(funcsMap, "toConstantCase", sr.ToConstantCase)
+	sprout.AddFunction(funcsMap, "toSnakeCase", sr.ToSnakeCase)
+	sprout.AddFunction(funcsMap, "toTitleCase", sr.ToTitleCase)
+	sprout.AddFunction(funcsMap, "untitle", sr.Untitle)
+	sprout.AddFunction(funcsMap, "swapCase", sr.SwapCase)
+	sprout.AddFunction(funcsMap, "split", sr.Split)
+	sprout.AddFunction(funcsMap, "splitn", sr.Splitn)
+	sprout.AddFunction(funcsMap, "substr", sr.Substring)
+	sprout.AddFunction(funcsMap, "indent", sr.Indent)
+	sprout.AddFunction(funcsMap, "nindent", sr.Nindent)
+	sprout.AddFunction(funcsMap, "seq", sr.Seq)
+}
