+++ registry/time/functions.go
@@ -4,24 +4,7 @@ import (
 	"strconv"
 	"strings"
 	"time"
-
-	"github.com/go-sprout/sprout"
 )
-
-// RegisterFunctions registers all functions of the registry.
-func (tr *TimeRegistry) RegisterFunctions(funcsMap sprout.FunctionMap) {
-	sprout.AddFunction(funcsMap, "date", tr.Date)
-	sprout.AddFunction(funcsMap, "dateInZone", tr.DateInZone)
-	sprout.AddFunction(funcsMap, "duration", tr.Duration)
-	sprout.AddFunction(funcsMap, "dateAgo", tr.DateAgo)
-	sprout.AddFunction(funcsMap, "now", tr.Now)
-	sprout.AddFunction(funcsMap, "unixEpoch", tr.UnixEpoch)
-	sprout.AddFunction(funcsMap, "dateModify", tr.DateModify)
-	sprout.AddFunction(funcsMap, "durationRound", tr.DurationRound)
-	sprout.AddFunction(funcsMap, "htmlDate", tr.HtmlDate)
-	sprout.AddFunction(funcsMap, "htmlDateInZone", tr.HtmlDateInZone)
-	sprout.AddFunction(funcsMap, "mustDateModify", tr.MustDateModify)
-}
 
 // Date formats a given date or current time into a specified format string.
 //
