+++ registry/time/time.go
@@ -20,3 +20,18 @@ func (tr *TimeRegistry) Uid() string {
 func (tr *TimeRegistry) LinkHandler(fh sprout.Handler) {
 	tr.handler = &fh
 }
+
+// RegisterFunctions registers all functions of the registry.
+func (tr *TimeRegistry) RegisterFunctions(funcsMap sprout.FunctionMap) {
+	sprout.AddFunction(funcsMap, "date", tr.Date)
+	sprout.AddFunction(funcsMap, "dateInZone", tr.DateInZone)
+	sprout.AddFunction(funcsMap, "duration", tr.Duration)
+	sprout.AddFunction(funcsMap, "dateAgo", tr.DateAgo)
+	sprout.AddFunction(funcsMap, "now", tr.Now)
+	sprout.AddFunction(funcsMap, "unixEpoch", tr.UnixEpoch)
+	sprout.AddFunction(funcsMap, "dateModify", tr.DateModify)
+	sprout.AddFunction(funcsMap, "durationRound", tr.DurationRound)
+	sprout.AddFunction(funcsMap, "htmlDate", tr.HtmlDate)
+	sprout.AddFunction(funcsMap, "htmlDateInZone", tr.HtmlDateInZone)
+	sprout.AddFunction(funcsMap, "mustDateModify", tr.MustDateModify)
+}
